OdfFormatReader - extended convert field

+++ ASCOfficeOdfFile/src/odf/odf_document.cpp
@@ -70,7 +70,7 @@ const std::wstring & odf_document::get_folder() const
     return impl_->get_folder();
 }
 
-long odf_document::get_office_mime_type() 
+int odf_document::get_office_mime_type() 
 {
 	return impl_->get_office_mime_type();
 }
